Split PaymentSheet and FlowController tests
